Search suggestions: boost recently visited repos/files/symbols
Previously, search suggestions didn't use personalized ranking. For
example, repos were only sorted by star count and didn't take into
account what repositories the user is visiting most frequently.

This PR reorders search suggestions so that recently visited
repos/files/symbols get ordered above results that the user has never
visited. This feature is enabled by the new `UserHistory` infrastructure
that was added for the fuzzy finder.
