+++ client/web/src/featureFlags/featureFlags.ts
@@ -17,6 +17,7 @@ export type FeatureFlagName =
     | 'enable-streaming-git-blame'
     | 'plg-enable-add-codehost-widget'
     | 'enable-rbac'
+    | 'user-history-ranking'
 
 interface OrgFlagOverride {
     orgID: string
