+++ client/web/src/nav/GlobalNavbar.tsx
@@ -8,6 +8,7 @@ import MagnifyIcon from 'mdi-react/MagnifyIcon'
 
 import { ContributableMenu } from '@sourcegraph/client-api'
 import { isErrorLike, isMacPlatform } from '@sourcegraph/common'
+import { UserHistory } from '@sourcegraph/shared/src/components/UserHistory'
 import { ExtensionsControllerProps } from '@sourcegraph/shared/src/extensions/controller'
 import { shortcutDisplayName } from '@sourcegraph/shared/src/keyboardShortcuts'
 import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
@@ -76,6 +77,7 @@ export interface GlobalNavbarProps
     showFeedbackModal: () => void
 
     setFuzzyFinderIsVisible: React.Dispatch<SetStateAction<boolean>>
+    userHistory?: UserHistory
 }
 
 /**
@@ -373,6 +375,7 @@ export const GlobalNavbar: React.FunctionComponent<React.PropsWithChildren<Globa
                         isSourcegraphDotCom={isSourcegraphDotCom}
                         searchContextsEnabled={searchContextsEnabled}
                         isRepositoryRelatedPage={isRepositoryRelatedPage}
+                        userHistory={props.userHistory}
                     />
                 </div>
             )}
