+++ client/web/src/search/home/SearchPageInput.tsx
@@ -9,6 +9,7 @@ import { SearchBox } from '@sourcegraph/search-ui'
 // The experimental search input should be shown on the search home page
 // eslint-disable-next-line  no-restricted-imports
 import { LazyCodeMirrorQueryInput } from '@sourcegraph/search-ui/src/experimental'
+import { UserHistory } from '@sourcegraph/shared/src/components/UserHistory'
 import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
 import { Settings } from '@sourcegraph/shared/src/schema/settings.schema'
 import {
@@ -60,6 +61,7 @@ interface Props
     autoFocus?: boolean
     queryState: QueryState
     setQueryState: (newState: QueryState) => void
+    userHistory?: UserHistory
 }
 
 const queryStateSelector = (
@@ -147,6 +149,7 @@ export const SearchPageInput: React.FunctionComponent<React.PropsWithChildren<Pr
             applySuggestionsOnEnter={applySuggestionsOnEnter}
             showSearchHistory={true}
             recentSearches={recentSearches}
+            userHistory={props.userHistory}
         />
     )
     return (
